+++ code/fabdem_loader.h
@@ -35,7 +35,7 @@
 class FabdemLoader : public TileLoader {
 public:
   // minLat and minLng name the SW corner of the tile, in degrees
-  virtual Tile *loadTile(const std::string &directory, float minLat, float minLng);
+  virtual Tile *loadTile(const std::string &directory, double minLat, double minLng);
 
 private:
 };
