+++ code/util.h
@@ -38,7 +38,7 @@ std::string trim(const std::string &s);
 
 // Adjust the given coordinate by an epsilon value away from 0, so that truncation
 // to int doesn't give incorrect values to due floating-point imprecision.
-float adjustCoordinate(float coordinate);
+double adjustCoordinate(double coordinate);
 
 /*
  * Split the given string by the given delimiter, putting the
